refactor(gateway): split GatewayTurnMixin into a facade + two topic siblings
gateway/run_turn.py held one 3,838-line class (93 methods) — past the ~2,000-line
split threshold in AGENTS.md. Mechanical extraction, no rewrites:

  gateway/run_turn.py        3884 -> 689 lines  (facade; 22 methods + nested types)
  gateway/run_turn_hmwa.py   new, 1641 lines     (39 _hmwa* methods)
  gateway/run_turn_exec.py   new, 1578 lines     (32 _run* methods)

GatewayTurnMixin now composes GatewayTurnHmwaMixin + GatewayTurnExecMixin (MRO
preserves every lookup, incl. class-level _STATUS_HINTS/_PreparedTurn/_HygienePlan
reached via self.*). Each sibling carries only the module-level imports its own
methods reference; logger keeps the same name ("gateway.run").

Evidence (worktree refactor/run-turn-split):
- 93/93 methods byte-identical to the pre-split class (AST source-segment equality)
- py_compile + import smoke green; gateway.run imports OK; MRO verified
- 35-file scoped suite: 423 passed / 4 failed — identical to the untouched live
  tree's 423/4, i.e. zero regressions (the 4 are pre-existing, unrelated)

Five tests asserted on `ast.parse(inspect.getsource(gateway.run_turn))`; they now
scan the module tuple (run_turn, run_turn_hmwa, run_turn_exec), matching the
existing run_turn_runner precedent in test_compression_session_id_persistence.
