+++ tests/gateway/test_compression_deferred_soft_result.py
@@ -24,6 +24,8 @@ import inspect
 from gateway import run as gateway_run
 from gateway import run_turn as gateway_run_turn
 from gateway import run_turn as gateway_run_turn
+from gateway import run_turn_exec as gateway_run_turn_exec
+from gateway import run_turn_hmwa as gateway_run_turn_hmwa
 
 
 def _calls(node: ast.AST) -> set[str]:
@@ -37,7 +39,8 @@ def _calls(node: ast.AST) -> set[str]:
 def _find_deferred_guarded_reset_chain() -> ast.If:
     """Return the ``if agent_result.get('compression_deferred') ... elif
     agent_result.get('compression_exhausted') ... reset_session`` chain."""
-    tree = ast.parse(inspect.getsource(gateway_run_turn))
+    tree = ast.parse("".join(inspect.getsource(m) for m in
+                            (gateway_run_turn, gateway_run_turn_hmwa, gateway_run_turn_exec)))
 
     for node in ast.walk(tree):
         if not isinstance(node, ast.If):
